src/requirements.txt
aiofiles==0.4.0
certifi==2019.11.28
chardet==3.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.25
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.9
mongoengine==0.19.1
multidict==4.7.5
pymongo==3.10.1
python-dotenv==0.12.0
rfc3986==1.3.2
sanic==19.12.2
six==1.14.0
sniffio==1.1.0
websockets==8.1
